+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## next [????.??.??]
+* Add `{e,E}limProxy` to `Data.Eliminator`.
+
 ## 0.8 [2021.03.12]
 * Require `singletons-base-3.0` and GHC 9.0.
 * Remove eliminators for `Data.Semigroup.Option`, which is deprecated as of